nvmet: separate our error classes
Separate out 'CFSError' and 'CFSNotFound' into a error.py to make
it available to other modules.
